backend/requirements.txt
flask
python-dotenv
psycopg2
supabase
jsonify
requests
openai